ResultTextCardを別ファイルに移動
