+++ encryption_sns/lib/process_view.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:encryption_sns/application_bloc.dart';
 import 'package:flutter/services.dart';
 import 'package:encryption_sns/Widget/ProcessView/bottom_navigationbar.dart';
+import 'package:encryption_sns/Widget/ProcessView/result_text_card.dart';
 
 class ProcessView extends StatefulWidget {
   final ApplicationBloc applicationBloc;
@@ -38,7 +39,7 @@ class _ProcessViewState extends State<ProcessView> {
             crossAxisAlignment: CrossAxisAlignment.center,
             children: <Widget>[
               _textField(),
-              _resultTextCard(),
+              ResultTextCard(applicationBloc: widget.applicationBloc),
             ],
           ),
         )
@@ -63,34 +64,6 @@ class _ProcessViewState extends State<ProcessView> {
         textAlign: TextAlign.left,
         focusNode: this.focusNode,
       ),
-    );
-  }
-
-  Widget _resultTextCard() {
-    return Card(
-      elevation: 4.0,
-      margin: const EdgeInsets.all(16.0),
-      child: Container(
-        padding: EdgeInsets.all(10),
-        width: 400,
-        height: 200,
-        child: StreamBuilder(
-          stream: widget.applicationBloc.showingText,                //出力用streamを取得
-          builder: (context, snapshot) {
-            return _resultText(snapshot.hasData ? snapshot.data.toString() : "");
-          },
-        ),
-      ),
-    );
-  }
-
-  Widget _resultText(String resultText) {
-    return GestureDetector(
-      child: new Tooltip(preferBelow: false,
-          message: "Copy", child: new Text(resultText)),
-      onTap: () {
-        Clipboard.setData(new ClipboardData(text: resultText));
-      },
     );
   }
 }
